Extract shared mouse UI coordinate conversion
Move top-left and Unity screen-space conversion into a focused helper so press and drag executors can share the same Game view coordinate contract without depending on the use case.
